+++ apps/deployer/lib/deployer/status/application.ex
@@ -223,7 +223,12 @@ defmodule Deployer.Status.Application do
       name: "deployex",
       sname: "deployex",
       node: Node.self(),
-      port: Application.get_env(:deployex_web, DeployexWeb.Endpoint)[:http][:port],
+      ports: [
+        %{
+          key: "PORT",
+          base: Application.get_env(:deployex_web, DeployexWeb.Endpoint)[:http][:port]
+        }
+      ],
       version: Application.spec(:deployer, :vsn) |> to_string,
       otp: check_otp_deployex.(),
       tls: Common.check_mtls(),
@@ -242,7 +247,7 @@ defmodule Deployer.Status.Application do
       crash_restart_count: crash_restart_count,
       force_restart_count: force_restart_count,
       start_time: start_time,
-      port: port
+      ports: ports
     } = Monitor.state(sname)
 
     check_otp_monitored_app = fn
@@ -260,7 +265,7 @@ defmodule Deployer.Status.Application do
       name: name,
       sname: sname,
       node: node,
-      port: port,
+      ports: ports,
       version: current_version(sname),
       otp: check_otp_monitored_app.(node, status),
       tls: Common.check_mtls(),